// GestionStock/src/app/services/openapi/model/utilisateur-request.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdresseRequest } from './adresse-request';


export interface UtilisateurRequest { 
    id?: number;
    nom: string;
    sexe: UtilisateurRequest.SexeEnum;
    login: string;
    numeroTel: string;
    email?: string;
    adresse: AdresseRequest;
}
export namespace UtilisateurRequest {
    export type SexeEnum = 'MASCULIN' | 'FEMININ';
    export const SexeEnum = {
        Masculin: 'MASCULIN' as SexeEnum,
        Feminin: 'FEMININ' as SexeEnum
    };
}