// GestionStock/src/app/services/openapi/model/utilisateur-response.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdresseResponse } from './adresse-response';


export interface UtilisateurResponse { 
    id?: number;
    nom?: string;
    login?: string;
    sexe?: UtilisateurResponse.SexeEnum;
    numeroTel?: string;
    email?: string;
    photo?: string;
    adresse?: AdresseResponse;
}
export namespace UtilisateurResponse {
    export type SexeEnum = 'MASCULIN' | 'FEMININ';
    export const SexeEnum = {
        Masculin: 'MASCULIN' as SexeEnum,
        Feminin: 'FEMININ' as SexeEnum
    };
}